Trying to implement 'streak' style texture function

+++ src/common/com_assets/com_perlin.h
@@ -1,6 +1,6 @@
 #ifndef PERLIN_H
 #define PERLIN_H
 
-extern double Perlin_Get2d(double x, double y, double freq, int depth);
+extern float Perlin_Get2d(float x, float y, float  freq, int depth);
 
 #endif  // PERLIN_H
